♻️ Build the pagefind bundle from the sitemap request, drop the CI step
Move generation into pagefindRoute.routemap: requesting /sitemap.xml builds
the bundle (if missing) by shelling out to pagefind.ts, waits, then enumerates
the files. Staticalize's own sitemap request now triggers and captures the
bundle in one pass, so the separate "Build Pagefind index" warmup step is gone
— the www deploy is back to serve → staticalize → deploy.

Generation's crawl re-requests /sitemap.xml; while a build is in flight
routemap returns [] without awaiting it, so the crawl driving the build can't
deadlock on itself. Deletes the pagefind plugin (its trigger now lives in the
route) and restores a `pagefind` task for building locally.

Verified against a fresh server: staticalize's first /sitemap.xml triggers the
build, waits, and captures all 232 bundle files (498 pages, 0 deadlocks).
