+++ www/deno.json
@@ -2,6 +2,7 @@
   "tasks": {
     "dev": "deno run -A @effectionx/watch deno run -A main.tsx",
     "staticalize": "deno run -A jsr:@frontside/staticalize@0.2.2/cli --site http://localhost:8000 --output=built --base=http://localhost:8000",
+    "pagefind": "deno run -A pagefind.ts",
     "test": "deno test --allow-run --allow-write --allow-read"
   },
   "lint": {
